Fix problems with pathfinder not working between core and integration
