+++ AssetBundle/AssetBundle.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 3804375300
+CRC: 3138239716
 AssetBundleManifest:
   AssetBundleInfos:
     Info_0:
