+++ AssetBundle/valheim-vehicles.manifest
@@ -1,15 +1,15 @@
 ManifestFileVersion: 0
-CRC: 2832432409
+CRC: 3227521597
 Hashes:
   AssetFileHash:
     serializedVersion: 2
-    Hash: 2998a6bb4570621bca55dfa635e94d40
+    Hash: 833da08f6e46e2e9e73589f2428317fb
   TypeTreeHash:
     serializedVersion: 2
     Hash: dcaa5347cbf726431e757fa742ed521b
   IncrementalBuildHash:
     serializedVersion: 2
-    Hash: 2998a6bb4570621bca55dfa635e94d40
+    Hash: 833da08f6e46e2e9e73589f2428317fb
 HashAppended: 0
 ClassTypes:
 - Class: 1
