+++ src/Eldritch.Valheim/Assets/Translations/English/eldritch-translations.json
@@ -1,0 +1,4 @@
+{
+  "eldritch_xeno_drone": "xeno-drone",
+  "eldritch_xeno_drone_desc": "A night terror. Hunts alone mosty but can appear in packs"
+}
