+++ src/Eldritch.Valheim/EldritchValheimPlugin.cs
@@ -27,104 +27,19 @@ public class EldritchValheimPlugin : BaseUnityPlugin
     LoggerProvider.Setup(Logger);
     LoggerProvider.LogDebug("Eldritchcore initialized");
 
+    // ReSharper disable once RedundantNameQualifier
+    Eldritch.Core.Nav.Pathfinding.Register(new ValheimPathfindingShim());
+
     InvokeRepeating(nameof(DebugCheckinMethod), 3f, 5f);
 
     PrefabManager.OnVanillaPrefabsAvailable += () =>
     {
-      LoadAssemblies();
+      PrefabRegistry.RegisterAllPrefabs();
     };
   }
 
   public void DebugCheckinMethod()
   {
     LoggerProvider.LogDebug("Checkin called");
-  }
-
-  public void InjectAlienPrefab()
-  {
-    if (assetBundle == null) return;
-    LoggerProvider.LogDebug($"Starting InjectAlienPrefab");
-    const string XenoAdultName = "Eldritch_XenoAdult";
-    const string assetName = "alientest";
-
-    var prefabAsset = assetBundle.LoadAsset<GameObject>(assetName);
-    if (prefabAsset == null) return;
-
-    var originalComponents = prefabAsset.GetComponents<Component>();
-    var clonedPrefab = PrefabManager.Instance.CreateClonedPrefab(XenoAdultName, prefabAsset);
-
-    clonedPrefab.AddComponent<ZNetView>();
-    clonedPrefab.AddComponent<XenoDroneSpawnHandler>();
-    var piece = clonedPrefab.AddComponent<Piece>();
-    piece.m_name = XenoAdultName;
-    piece.m_icon = Sprites.GetSprite("anchor");
-
-    var components = clonedPrefab.GetComponents<Component>();
-    foreach (var component in components)
-    {
-      if (component == null)
-      {
-        LoggerProvider.LogError("Got null component");
-        continue;
-      }
-      LoggerProvider.LogDebug(component.GetType().ToString());
-    }
-
-    foreach (var component in originalComponents)
-    {
-      if (component == null)
-      {
-        LoggerProvider.LogError("Got null component");
-        continue;
-      }
-      LoggerProvider.LogDebug(component.GetType().ToString());
-    }
-
-    PieceManager.Instance.AddPiece(new CustomPiece(clonedPrefab, true,
-      new PieceConfig
-      {
-        PieceTable = "Hammer", // for now.
-        Name = "XENO",
-        Description = "Xenos....they're everywhere...",
-        Icon = Sprites.GetSprite("anchor"), // for testing.
-        Category = "Eldritch",
-        Enabled = true,
-        Requirements =
-        [
-          new RequirementConfig
-          {
-            Amount = 1,
-            Item = "Wood",
-            Recover = true
-          }
-        ]
-      }));
-
-    LoggerProvider.LogDebug($"Added {piece}");
-  }
-
-  /// <summary>
-  /// Loads from eldritch core. (this might need to be done here so eldritch.Core gameobjects are resolved with eldritch core scripts)
-  /// </summary>
-  public void LoadAssemblies()
-  {
-    if (!assetBundle)
-    {
-      assetBundle = Entry.LoadAssembly();
-    }
-
-    if (assetBundle == null) return;
-
-    Sprites = assetBundle.LoadAsset<SpriteAtlas>("icons");
-
-    try
-    {
-
-      InjectAlienPrefab();
-    }
-    catch (Exception e)
-    {
-      LoggerProvider.LogError($"Failed to load alien prefab {e}");
-    }
   }
 }
