+++ src/ValheimRAFT.Unity/Assets/ValheimVehicles/ExternalAssets/Enemies/xeno/XenoAnimatorController.controller
@@ -1446,7 +1446,7 @@ BlendTree:
     m_Motion: {fileID: 17259882852881034, guid: 595cee29b9912f04d9294bb4511d96bb, type: 3}
     m_Threshold: 0.25
     m_Position: {x: 0, y: 0}
-    m_TimeScale: 0.1
+    m_TimeScale: 0.5
     m_CycleOffset: 0
     m_DirectBlendParameter: Blend
     m_Mirror: 0
