+++ src/ValheimVehicles/ValheimVehicles.Controllers/WaterZoneController.cs
@@ -13,6 +13,7 @@ using ValheimVehicles.SharedScripts;
 using ValheimVehicles.Controllers;
 using ValheimVehicles.Shared.Constants;
 using ValheimVehicles.Structs;
+using Zolantris.Shared;
 using Logger = Jotunn.Logger;
 
 namespace ValheimVehicles.Components;
